+++ plugins/openai/src/stt.ts
@@ -27,6 +27,7 @@ export class STT extends stt.STT {
   #opts: STTOptions;
   #client: OpenAI;
   label = 'openai.STT';
+  private abortController = new AbortController();
 
   /**
    * Create a new instance of OpenAI STT.
@@ -145,13 +146,19 @@ export class STT extends stt.STT {
     const config = this.#sanitizeOptions(language);
     buffer = mergeFrames(buffer);
     const file = new File([this.#createWav(buffer)], 'audio.wav', { type: 'audio/wav' });
-    const resp = await this.#client.audio.transcriptions.create({
-      file,
-      model: this.#opts.model,
-      language: config.language,
-      prompt: config.prompt,
-      response_format: 'json',
-    });
+
+    const resp = await this.#client.audio.transcriptions.create(
+      {
+        file,
+        model: this.#opts.model,
+        language: config.language,
+        prompt: config.prompt,
+        response_format: 'json',
+      },
+      {
+        signal: this.abortController.signal,
+      },
+    );
 
     return {
       type: stt.SpeechEventType.FINAL_TRANSCRIPT,
@@ -170,5 +177,9 @@ export class STT extends stt.STT {
   /** This method throws an error; streaming is unsupported on OpenAI STT. */
   stream(): stt.SpeechStream {
     throw new Error('Streaming is not supported on OpenAI STT');
+  }
+
+  async close(): Promise<void> {
+    this.abortController.abort();
   }
 }